requirements.txt
streamlit>=1.35.0
yt-dlp>=2024.11.18
pydub>=0.25.1
audioop-lts>=0.2.1  # pydub compatibility shim for Python 3.13+
openai>=1.50.0
anthropic>=0.40.0
fpdf2>=2.7.9
Pillow>=10.0.0
python-dotenv>=1.0.0
